+++ griptape/drivers/file_manager/base_file_manager_driver.py
@@ -1,11 +1,11 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
+from typing import Optional
 
-from attrs import Factory, define, field
+from attrs import define, field
 
-import griptape.loaders as loaders
-from griptape.artifacts import BaseArtifact, ErrorArtifact, InfoArtifact, ListArtifact, TextArtifact
+from griptape.artifacts import BlobArtifact, InfoArtifact, TextArtifact
 
 
 @define
@@ -17,57 +17,28 @@ class BaseFileManagerDriver(ABC):
         loaders: Dictionary of file extension specific loaders to use for loading file contents into artifacts.
     """
 
-    default_loader: loaders.BaseLoader = field(default=Factory(lambda: loaders.BlobLoader()), kw_only=True)
-    loaders: dict[str, loaders.BaseLoader] = field(
-        default=Factory(
-            lambda: {
-                "pdf": loaders.PdfLoader(),
-                "csv": loaders.CsvLoader(),
-                "txt": loaders.TextLoader(),
-                "html": loaders.TextLoader(),
-                "json": loaders.TextLoader(),
-                "yaml": loaders.TextLoader(),
-                "xml": loaders.TextLoader(),
-                "png": loaders.ImageLoader(),
-                "jpg": loaders.ImageLoader(),
-                "jpeg": loaders.ImageLoader(),
-                "webp": loaders.ImageLoader(),
-                "gif": loaders.ImageLoader(),
-                "bmp": loaders.ImageLoader(),
-                "tiff": loaders.ImageLoader(),
-            },
-        ),
-        kw_only=True,
-    )
+    workdir: str = field(kw_only=True)
+    encoding: Optional[str] = field(default=None, kw_only=True)
 
-    def list_files(self, path: str) -> TextArtifact | ErrorArtifact:
+    def list_files(self, path: str) -> TextArtifact:
         entries = self.try_list_files(path)
         return TextArtifact("\n".join(list(entries)))
 
     @abstractmethod
     def try_list_files(self, path: str) -> list[str]: ...
 
-    def load_file(self, path: str) -> BaseArtifact:
-        extension = path.split(".")[-1]
-        loader = self.loaders.get(extension) or self.default_loader
-        source = self.try_load_file(path)
-        result = loader.load(source)
-
-        if isinstance(result, BaseArtifact):
-            return result
+    def load_file(self, path: str) -> BlobArtifact | TextArtifact:
+        if self.encoding is None:
+            return BlobArtifact(self.try_load_file(path))
         else:
-            return ListArtifact(result)
+            return TextArtifact(self.try_load_file(path).decode(encoding=self.encoding), encoding=self.encoding)
 
     @abstractmethod
     def try_load_file(self, path: str) -> bytes: ...
 
     def save_file(self, path: str, value: bytes | str) -> InfoArtifact:
-        extension = path.split(".")[-1]
-        loader = self.loaders.get(extension) or self.default_loader
-        encoding = None if loader is None else loader.encoding
-
         if isinstance(value, str):
-            value = value.encode() if encoding is None else value.encode(encoding=encoding)
+            value = value.encode() if self.encoding is None else value.encode(encoding=self.encoding)
         elif isinstance(value, (bytearray, memoryview)):
             raise ValueError(f"Unsupported type: {type(value)}")