+++ griptape/loaders/__init__.py
@@ -1,26 +1,28 @@
 from .base_loader import BaseLoader
-from .base_text_loader import BaseTextLoader
+from .base_file_loader import BaseFileLoader
+
 from .text_loader import TextLoader
 from .pdf_loader import PdfLoader
 from .web_loader import WebLoader
 from .sql_loader import SqlLoader
 from .csv_loader import CsvLoader
-from .dataframe_loader import DataFrameLoader
 from .email_loader import EmailLoader
-from .image_loader import ImageLoader
-from .audio_loader import AudioLoader
+
 from .blob_loader import BlobLoader
+
+from .image_loader import ImageLoader
+
+from .audio_loader import AudioLoader
 
 
 __all__ = [
     "BaseLoader",
-    "BaseTextLoader",
+    "BaseFileLoader",
     "TextLoader",
     "PdfLoader",
     "WebLoader",
     "SqlLoader",
     "CsvLoader",
-    "DataFrameLoader",
     "EmailLoader",
     "ImageLoader",
     "AudioLoader",