+++ tests/unit/drivers/vector/test_local_vector_store_driver.py
@@ -3,10 +3,10 @@ import pytest
 from griptape.artifacts import TextArtifact
 from griptape.drivers import LocalVectorStoreDriver
 from tests.mocks.mock_embedding_driver import MockEmbeddingDriver
-from tests.unit.drivers.vector.test_base_local_vector_store_driver import BaseLocalVectorStoreDriver
+from tests.unit.drivers.vector.test_base_vector_store_driver import TestBaseVectorStoreDriver
 
 
-class TestLocalVectorStoreDriver(BaseLocalVectorStoreDriver):
+class TestLocalVectorStoreDriver(TestBaseVectorStoreDriver):
     @pytest.fixture()
     def driver(self):
         return LocalVectorStoreDriver(embedding_driver=MockEmbeddingDriver())